Backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
langchain-ollama
python-multipart
pypdf2
python-docx
pydantic
openai
python-dotenv
requests
